Se agregan las funcionalidades relacionadas con la captura y calculo de las preguntas zero, incluyendo (vistas, procedimientos, validaciones y pruebas)

+++ EwoQ/Models/ReporteIncidentesViewModel.cs
@@ -271,7 +271,7 @@ namespace EwoQ.Models
         public string ListAfte { get; set; }
         public string ListPlan { get; set; }
         public string List4M { get; set; }
-        public int MaxMId { get; set; }
+        public long MaxMId { get; set; }
 
         #endregion
     }
